Remove old git log dependency
